remove bogus timeout delay
This was an ill-conceived attempt to keep graphs updated in sync, but it no
longer worked at all for two reasons: it only ever worked for instrumentations
loaded when the page loads (which is no longer any of them), and it was defeated
by the fact that we update graphs when we receive backend data.  This change was
necessary to make features work that involve redrawing the graph based on
server-side data after the user interacts with it (e.g., selects a component in
a heat map).